requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-SocketIO==4.3.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
keyboard==0.13.5
MarkupSafe==1.1.1
mouse==0.7.1
Pillow==7.2.0
PyMsgBox==1.0.8
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
Desktopmagic~=14.3.11
playsound~=1.2.2